refactor(api-client): centralize run lifecycle
